requirements.txt
pandas
py-AutoClean
ydata-profiling
pdfkit
matplotlit
pgwalker
streamlit
